Added configurable window space
This means that the user can specify an area to be used for tiling,
which the window manager tries to use, but can shrink if necessary.
This allows for bars and window-free areas on the user's desktop.
However, offsets still need to be implemented, and added to the
ScreenSize struct.