Rewrote window property lookup to fix #16
Instead of assuming everything is to be handled unless it has window
type `DOCK`, we now explicitly handle windows that have type `NORMAL`.
However, we also assume that a window which doesn't set it's type (such
is the case with `mpv`, `sxiv` and others) is to be managed as well.