Added source annoations for future directions
Probably we'll have to do some cleanup regarding the use of `Vec<Tag>`
everywhere, as we don't want any ordering. `HashSet<Tag>` would be a
good fit, if it weren't for some places where we would use it as a key
to a `HashTable`, which is possible, but ridiculously slow for
implementations I can imagine.