Implemented tagset-based tag stack overlap for multiple screens (addresses #26)
Now we get information on when two tagsets shown simultaneously overlap.
This is very useful since you probably are aware of the tags that are
*present* on a tag stack's top entry, but you can't easily know which of
those are actually shown. You have a guarantee, however, that they are
somewhere on screen, but they could have been claimed by a screen with a
lower index already. This information we now compute and output. This
also goes hand-in-hand with other changes around the logic of tagset
treatment. We need to be *very* careful to not blur the boundary between
*visible* and *current* tags on each screen.

This probably means that this implementation is buggy and has
unexpectedly rough corners in some places. For simple usecases, no bugs
are expected, however. Time to get fuzzing though.